sitecar/requirements.txt
Django==1.9.4
dj-database-url==0.4.1
django-chartit==0.1
django-chartit2==0.2.2
django-haystack==2.4.1
elasticsearch==2.3.0
gunicorn==19.6.0
lxml==3.6.0
psycopg2==2.6.1
simplejson==3.8.2
urllib3==1.15.1
whitenoise==3.2
